# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "مستندات API سامانه لد ",
        "description": "توضیحات نحوه استفاده از APIها",
        "contact": {
            "email": "hreza779@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "سرور اصلی API"
        }
    ],
    "paths": {
...
        "/api/exams": {
            "get": {
                "tags": [
                    "Exams"
                ],
                "summary": "Get all exams",
                "operationId": "a54e05ebd2b23c9f2329a2a483272bc3",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "draft",
                                "archived"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of exams"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exams"
                ],
                "summary": "Create a new exam",
                "operationId": "fa2de658c3313746441306155597563e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "duration",
                                    "passing_score"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Math Exam"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Basic math questions"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "passing_score": {
                                        "type": "integer",
                                        "example": 70
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "draft",
                                            "archived"
                                        ],
                                        "example": "draft"
                                    },
                                    "question_ids": {
                                        "description": "Array of question IDs to attach",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exam created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exams/{exam}": {
            "get": {
                "tags": [
                    "Exams"
                ],
                "summary": "Get a specific exam",
                "operationId": "112c80bb353c3bb00ddf36a06ea617cf",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam details"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Exams"
                ],
                "summary": "Update an exam",
                "operationId": "ff1b5cc2782780e37965d541920e15cd",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "integer"
                                    },
                                    "passing_score": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "draft",
                                            "archived"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam updated successfully"
                    },
                    "404": {
                        "description": "Exam not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Exams"
                ],
                "summary": "Delete an exam",
                "operationId": "2749cce3d76ee23f433dc1ae5cc2bce0",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Exam deleted successfully"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exams/{exam}/questions": {
            "post": {
                "tags": [
                    "Exams"
                ],
                "summary": "Attach questions to exam",
                "operationId": "ce7741760a7ea5302ae889ec411ff4f5",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "questions"
                                ],
                                "properties": {
                                    "questions": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer"
                                                },
                                                "order": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "question_id": 1,
                                                "order": 1
                                            },
                                            {
                                                "question_id": 2,
                                                "order": 2
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Questions attached successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exams/{exam}/questions/{question}": {
            "delete": {
                "tags": [
                    "Exams"
                ],
                "summary": "Detach question from exam",
                "operationId": "eb8c172e1bdb14255177f0a4fc0ab0a9",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question detached successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}